+++ src/components/manage/Widgets/DatetimeWidget.jsx
@@ -203,6 +203,7 @@ class DatetimeWidget extends Component {
           >
             <SingleDatePicker
               date={datetime}
+              disabled={this.props.isDisabled}
               onDateChange={this.onDateChange}
               focused={focused}
               numberOfMonths={1}
@@ -223,6 +224,7 @@ class DatetimeWidget extends Component {
               })}
             >
               <TimePicker
+                disabled={this.props.isDisabled}
                 defaultValue={datetime}
                 value={datetime}
                 onChange={this.onTimeChange}
@@ -238,7 +240,7 @@ class DatetimeWidget extends Component {
             </div>
           )}
           <button
-            disabled={!datetime}
+            disabled={this.props.isDisabled || !datetime}
             onClick={() => this.onResetDates()}
             className="item ui noborder button"
           >