+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Feature
 
+- Added Schema Editor within Dexterity Content-Types Controlpanel @rexalex @avoinea #1517
+- Added Blocks Layout Editor within Dexterity Content-Types Controlpanel @avoinea #1517
+- Added missing components for Email and Url widgets #1246 @rexalex
+
 ### Bugfix
 
 ### Internal
@@ -315,6 +319,7 @@
 
 ### Breaking
 
+### Feature
 
 ### Bugfix